**New insertion modes added (afterbody and afterafterbody)
**In body modified to display new tags (b, i, p) 

**Text strings are appended to the body node (eg 'abc' produces the
minimal dom with 'abc' inside the body element).
 
**Text string appended to b, i, p elements (eg. <p>qwerty produces the
minimal dom with <p>qwerty</p> insode the body).
 
**New class (StackUpdater) created to manage the stack of open elements
and generate the DOM.

**Changes to all insertion modes to use the StackUpdater class.

!!! Note that for the b, i, p tags now it is only displaying the nodes
and nesting them if they are in order but it is not complete, there are
algorithms for formatting elements. In case someone can update it, check
the InBody class, I put the link to those missing steps.